Final Project on Telecom Churn
The purpose of this project is to analyze the factors/variables impacting churn from historical data and construct classification-based prediction models for predicting customer churn in the future, which will assist the telecom firm in planning for improved services. To predict churn customers in the dataset, Data Mining methods are used.